Update time_log.py
Hey Andrew,
This is the propose to change the overlap from owners to EMPLOYEE, so it doesn't conflict if someone add a new timelog at the same time. Unless it is the same user.
And apparently this code has been fixed about the time starting and finishing at the same time. e.g finishing one task at 10:00:00 and starting the other one at 10:00:00 as well.