Outstanding amount exceeds on decimal places.
The outstanding amount exceeds the decimal place if the Total amount is partially paid. This update will temporarily fix the decimal issue of the Payment tool when it gets the Outstanding amount from Purchase Invoice. .toFixed(2) is a javascript code that will round off float values to decimal number of 2 or .00 which is the standard currency format.