Added Item Manufacturer, Item Manufacturer Alias and Item Manufacturer Part No instead of Data fields
I have changed the data fields to links which is important for
manufacturing business to keep the manufacturers and manufacturer part
numbers consistent.

To be checked by Frappe before merge:

- [ ] Please make the Manufacturer Part Number selection dependant on
the Manufacturer that the user has selected.
- [ ] If user clicks "New Manufacturer Part Number", could you read the
part number that was selected before?
- [ ] Can we show the associated Item Manufacturer Aliase's in the Item
Manufactuerer page or is this only for reports?
- [ ] Can we show the associated Item Manufacturer Part No in the Item
Manufactuer page or again reporting?