Fix calendar view for Timesheet
All the calendar events for Timesheets show up on one date. This is because there was a change to the timesheet_calendar.js file when the Gantt chart was added, but the get_events for the calendar wasn't fixed at the same time. Just fixing the reference to the start and end times.