Update issue.js
Make the close / reopen button more flexible for customized user status.

For example, if the user has several status, it will now show the Close button every time the ticket is not closed! It will show Reopen only, if the ticket is closed, and not if it's on hold for example.

This will make it more flexible.