Fix total monthly sales
We need to commit the db queries, before we call methods which update other doctypes, which rely on SQL queries, because otherwise the calculation of them is wrong and doesn't consider the current submitted or cancelled sales invoice.

I only saw problems with the current_month_sales. However, update_project relies on a direct SQL query as well. I don't know if update_time_sheet is affected, but it doesn't do any harm to commit the database query before calling the other methods.