Search for item_code (#13912)
Item query can be called from custom script to search for an additional field. In this case the item_name, item_group and item_description are still searchable, but not item_code. As the item code is a must have field in search results, I have added this to the non-changable search fields in this query.