Serial Number validation in stock entry
Stock Entry allows  same Serial Number items (duplication of serial no). Even though the product is sold (contain Delivery Note),again the  another product with same serial number is given it generate the new product with previous serial_no.In this Serial No list the creation document no is rewritten. So it allows repeation/duplication of serial number.In order to avoid the duplication the file is modified.