feat: Relational UOM Query
Make the list of UOMs that can be selected during a transaction dependent on the respective item. In order to be selectable in the link field a UOM either has to be the stock UOM of the item, be in the list of UOM conversion details or has a set conversion factor relative to the stock UOM of the item. Furthermore, the respective conversion factor is shown as preview in the link field.