feat: inform users about two extra chars on amending
GST rules mandates 16 as the upper limit of number of characters. Users therefore might unknowingly set 15/16 chars in the naming series which will cause issue upon amending(extra 2 chars). The mess it can create in the middle of a financial year is huge considering the fact that we have immutable ledgers.

This PR informs users to limit it to 14 in case of submittable documents.