feat(india): E Invoice Eway Bill Distance is calculated automatically.
Passing Distance as ‘0’ is an indication that the user wants the eway bill system to take the distance as per the E Way Bill database.

I have added description to notify the user about this and added functionality to save that distance in Sales Invoice Distance Field & show the auto calculated distance.