fix: `payment entry is already created` when making mpesa express payment request from posawesome
POSawesome creates sales invoices of type `is_pos`. Due to this, they are not handled as `POS Invoice` in the get_amount method in payment_request.
These are cash invoices hence no outstanding amount. This leads to the `frappe.throw(_("Payment Entry is already created"))` block when making payment requests.

fixes: https://github.com/frappe/erpnext/issues/34598