Fix: duplicate customer creation
The old `party_exists()` method uses the `frappe.db.get_value()`  to fetch a contact name using `current user` which causes it to skip validating other contact entries linked to the same `user email` that results in creating duplicate entries.

- Use `frappe.db.get_all( )` instead to grab all contacts related to the user email and loop through the result and return a Dict of `party_exist boolean and `contact` that has empty doctype links

The old  `create_party_contact()` method was adding new entries to contacts without checking if there is a contact with empty links related to the user.
- Use the `contact` key that we get from the new `party_exists()` method to determine whether to edit and use an existing contact or add a new one.