Change default Customer naming to Naming Series
On a new installation, a customer is identified by their name. This becomes a problem if you get customers who share names. I think the default naming scheme should be series, and then an administrator can change it to the customer's name if need be as opposed to it being the other way around.