Fix: CSS not applied to product title
The product_title css is not applied in the list view of the product UI because the class in the link is null and inline style statement is applied instead. 

This commit removes the null css class and inline style statement, then enclosing it in a div tag with the product_title css class applied, as it is in the grid view (see grid.js).