Bug Fixes : False Job card overlap Error
Version Using : 14.47.4
Bug found: False Job card overlap error.
Details: Consider the example below.
A job card with time log 2023-11-22 10:25:00 to 2023-11-22 11:40:00 is throwing error while saving. 
So as per the existing algorithm for job cards I found below time logs which are overlapping
1. 2023-11-22 10:20:00 to 2023-11-22 10:30:00
2. 2023-11-22 10:30:00 to 2023-11-22 10:40:00
3. 2023-11-22 10:40:00 to 2023-11-22 10:45:00
4. 2023-11-22 10:45:00 to 2023-11-22 10:50:00
and Production capacity is 3.
Now if you observe closely  only 1 job card is active at any instance of time. so technically all these 4 job cards we can grouped as 1 and can have more job cards as we have capacity 3.

So the changes we made as below:
we reach to the comparison point as per the previous algorithm and then we have introduced a new layer for finding set of job cards which are not overlapping(sequential). Considering those group job cards as single overlapping set for other job cards so comparison will be correct in this case.