Update report date assignment in process_statement_of_accounts
The code for assigning the report date in the process of statement of accounts file has been modified. Previously, it always pulled the "to_date" regardless of the report. Now, it checks if the report is a "General Ledger" and takes "to_date"; otherwise, it will take the "posting_date" as the report date.