Add bank account name to payment entry dict
In the bank reconciliation tool, the bank account name is now being retrieved and added to the payment entry dictionary. This was done by calling frappe.get_value for the "Bank Account" Doctype to get its name and then adding it to the payment entry dictionary. This will add more clarity to the payment records in the system.